fix(hipsparse): widen testhyb::ell_nnz to int64_t to match rocSPARSE (AISPARSE-681)
The hipSPARSE test harness reinterprets the opaque rocSPARSE HYB handle as its
own struct testhyb. rocSPARSE widened _rocsparse_hyb_mat::ell_nnz to int64_t so
that ell_width * m cannot overflow; the 32-bit mirror field left the struct
layout misaligned, so all fields after it were misread (garbage ell_nnz caused
std::vector allocation failures in hybmv). Mirror the int64_t width and adjust
the two 32-bit readers (explicit casts / matching unit_check type).

+++ projects/hipsparse/clients/include/utility.hpp
@@ -7634,12 +7634,15 @@ double get_time_us_sync(hipStream_t stream);
 }
 #endif
 
+// Mirror of _rocsparse_hyb_mat's layout. ell_nnz must be int64_t here to match
+// rocSPARSE, where it was widened so that ell_width * m cannot overflow when
+// sizing the ELL allocations; a 32-bit field would desynchronize the layout.
 struct testhyb
 {
     int                     m;
     int                     n;
     hipsparseHybPartition_t partition;
-    int                     ell_nnz;
+    int64_t                 ell_nnz;
     int                     ell_width;
     int*                    ell_col_ind;
     void*                   ell_val;
